+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 (Add entries below in the order of server, console, cli, docs, others)
 
+## v2.0.8
+
 - server: fix nullability of object relationships (close #7201)
-- server: optimize SQL query generation with LIMITs (close #5745)
 - server: update non-existent event trigger, action and query collection error msgs (close #7396)
 - server: fix broken `untrack_function` for non-default source
 - server: Adding support for TLS allowlist by domain and service id (port)